chore(deps): bump wafer-run to 21446515 (W3-W5, WR-39) and adapt
Moves the pin to wafer-run main 2144651539b55281e341e1a90e01b43bb7d8afa3
(WR-10..WR-18 and WR-39) and makes the changes its API requires:

- WR-13: ListOptions.limit is Option<u32>. Every literal becomes Some(n)
  and the repo page-size parameters that feed it are u32, so a zero or
  negative size can no longer mean "every row" by accident.
- WR-12: the three CryptoService impls drop master sign/verify; the
  browser and Cloudflare tests call sign_for, and an unknown hash scheme
  is MalformedHash.
- WR-11: the three LoggerService impls take the caller and render one
  line through the shared log_line::LogLine (caller=… msg=… fields).
  ResourceType::Auth gets its prepared-plan spelling.
- WR-14: the Cloudflare and browser DbExec row primitives take the
  statement's JsonColumns; their decode tests now pin that only a JSON
  column's text is parsed.
- WR-15: SqliteVecService::new is fallible.
- WR-39: the browser vector service implements rename_index — the
  index's tables through a staging name, then its registry row, in one
  transaction, refusing a case twin — with its statements run on real
  SQLite in native tests.

+++ crates/impresspress-core/src/blocks/admin/pages/network.rs
@@ -270,7 +270,7 @@ pub async fn network_inbound_detail(ctx: &dyn Context, msg: &Message) -> OutputS
     let method = msg.query("method").to_string();
     let path = msg.query("path").to_string();
     let offset: i64 = msg.query("offset").parse().unwrap_or(0);
-    let limit: i64 = 20;
+    let limit: u32 = 20;
 
     // One more than the page shows, to learn whether a next page exists.
     let rows = match request_logs::list_for_path(ctx, &method, &path, offset, limit + 1).await {
@@ -278,7 +278,7 @@ pub async fn network_inbound_detail(ctx: &dyn Context, msg: &Message) -> OutputS
         Err(e) => return crate::blocks::crud::db_error_internal(e, "Network inbound detail"),
     };
 
-    let has_more = rows.len() as i64 > limit;
+    let has_more = rows.len() > limit as usize;
     let display_rows = if has_more {
         &rows[..limit as usize]
     } else {
@@ -306,7 +306,7 @@ pub async fn network_inbound_detail(ctx: &dyn Context, msg: &Message) -> OutputS
             html! { p .text-center .text-muted { "No requests logged for this path" } },
         ))
         @if has_more {
-            @let next_offset = offset + limit;
+            @let next_offset = offset + i64::from(limit);
             div .text-center .p-2 {
                 button .btn .btn--secondary .btn--sm
                     hx-get={"/b/admin/network/detail/inbound?method=" (method) "&path=" (path) "&offset=" (next_offset)}
